Fixed HRDReportFactory attributes indentation
* Reinstated and reverted S3 services that fixed S3LambdaTests
* Refactored Lambda SQS message processing into helper module
  so that it can be used inter-package or module
* Reverted libiap to `0.3.0`
  Note: ICA automation work and pipeline branch shall take
  care of this related deps upgrade there.
